realocação para melhor manutençao
