+++ lib/Src/views/adm/studantDetails/pages/page_studant_anddress.dart
@@ -3,10 +3,8 @@
 import 'package:cefops/Shared/themes/app_textstayle.dart';
 import 'package:cefops/Src/controller/controller_cep.dart';
 import 'package:cefops/Src/controller/studants/studant_all_info_controller.dart';
-import 'package:cefops/Src/controller/studants/studant_anddress_controller.dart';
 import 'package:cefops/Src/repository/cepAuto/cep_repository.dart';
-import 'package:cefops/Src/views/studantDetails/controller/controller_studantDetails.dart';
-import 'package:cefops/Src/views/studantDetails/widget/widget_form_studantDetails.dart';
+import 'package:cefops/Src/views/adm/studantDetails/widget/widget_form_studantdetails.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
