+++ lib/Src/views/adm/studantDetails/widget/widget_dropdown_gender.dart
@@ -1,7 +1,5 @@
-
-import 'package:cefops/Src/controller/status.dart';
 import 'package:cefops/Src/controller/studants/studant_info_controller.dart';
-import 'package:cefops/Src/views/studantDetails/controller/controller_studantDetails.dart';
+import 'package:cefops/Src/views/adm/studantDetails/controller/controller_studantDetails.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
