+++ lib/Src/views/adm/studantDetails/widget/widget_dropdown_permission.dart
@@ -1,5 +1,5 @@
 import 'package:cefops/Shared/Security/Controller/userController.dart';
-import 'package:cefops/Src/views/studantDetails/controller/user_controller.dart';
+import 'package:cefops/Src/views/adm/studantDetails/controller/user_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get/get_state_manager/src/rx_flutter/rx_obx_widget.dart';
