+++ lib/Src/views/adm/studantDetails/widget/widget_dropdown_state.dart
@@ -1,11 +1,9 @@
-
-import 'package:cefops/Src/controller/studants/studant_info_controller.dart';
-import 'package:cefops/Src/views/studantDetails/controller/controller_studantDetails.dart';
+import 'package:cefops/Src/views/adm/studantDetails/controller/controller_studantDetails.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-class DropDownCivilState extends StatelessWidget {
-  const DropDownCivilState({Key? key}) : super(key: key);
+class DropDownStatus extends StatelessWidget {
+  const DropDownStatus({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -13,7 +11,7 @@ class DropDownCivilState extends StatelessWidget {
     return Obx(
             () {
           return DropdownButton<String>(
-            value: controller.civilState.value,
+            value: controller.status.value,
             icon: const Icon(Icons.arrow_downward),
             iconSize: 24,
             elevation: 16,
@@ -26,11 +24,10 @@ class DropDownCivilState extends StatelessWidget {
             ),
             onChanged: (String? newValue) {
 
-              controller.civilState.value = newValue!;
-              StudantInfoController.data.stsCivil.value=newValue;
+              controller.status.value = newValue!;
 
             },
-            items: <String>['Estado Civil','Solteiro', 'Casado','Viúvo','Divorciado','Separado',]
+            items: <String>['Ativo', 'Inativo',]
                 .map<DropdownMenuItem<String>>((String value) {
               return DropdownMenuItem<String>(
                 value: value,
