+++ lib/Src/views/adm/studantDetails/widget/widget_setstudant_details_routes.dart
@@ -1,10 +1,8 @@
-
 import 'package:cefops/Src/controller/studants/studant_all_info_controller.dart';
-import 'package:cefops/Src/views/studantDetails/pages/page_forms_studants_documents.dart';
-import 'package:cefops/Src/views/studantDetails/pages/page_studant_anddress.dart';
-import 'package:cefops/Src/views/studantDetails/pages/page_user_infos.dart';
+import 'package:cefops/Src/views/adm/studantDetails/pages/page_forms_studants_documents.dart';
+import 'package:cefops/Src/views/adm/studantDetails/pages/page_studant_anddress.dart';
+import 'package:cefops/Src/views/adm/studantDetails/pages/page_user_infos.dart';
 import 'package:cefops/Src/views/studants/requeriment/page_my_requeriment.dart';
-
 import '../pages/page_forms_studants.dart';
 
 SetStudantDetails(int pos,{required bool fromList1}) {
