+++ lib/Src/views/adm/studantDetails/widget/widget_studant_infos_routes.dart
@@ -1,0 +1,16 @@
+import 'package:cefops/Src/views/adm/studantDetails/pages/page_studant_infos_submenu.dart';
+import 'package:cefops/Src/views/adm/studantDetails/pages/page_studant_notes.dart';
+import 'package:cefops/Src/views/adm/studantDetails/pages/page_studant_requeriment.dart';
+
+SetStudantPage(int pos,{required String id}) {
+  switch (pos) {
+    case 0:
+      return ViewInfoFromStudant();
+    case 1:
+      return ViewInfoNotesStudant();
+    case 5:
+      return StuddantRequeriment(id);
+
+
+  }
+}
